feat(front): create home page
