+++ client/src/pages/Home/Home.scss
@@ -1,0 +1,20 @@
+.home {
+  .pages-container {
+    margin-top: 1rem;
+    div {
+      border-bottom: 1px solid #e2e8f0;
+      padding: 0.25rem;
+      display: flex;
+      justify-content: space-between;
+      cursor: pointer;
+      
+      &:hover {
+        background-color: #f7fafc;
+      }
+
+      p {
+        color: #999;
+      }
+    }
+  }
+}
