+++ client/src/styles.scss
@@ -30,6 +30,3 @@ a {
   }
 }
 
-li {
-  display: list-item;
-}
